fix the equality check
CodeQL:
The method signature of Equals takes an object as an argument. It is therefore common to test whether the parameter is of the same type as the object that Equals is being called on. This should not be done using as and then checking for null because this technique does not limit the argument to being the exact same type as the object. It will not return null if the argument is a subtype of the object's type.

As an exception to the rule, it is acceptable to use as when your class is sealed, because the concern about subclassing does not then apply. That said, it is probably best to avoid it even then, because sealed classes can sometimes become unsealed later - it's sensible to try and avoid potential problems down the line.